Efetuando requisicao para a api externa em caso de pagamento por cartao

+++ javascript/smallProjects/node/payfast/routes/pagamentos.js
@@ -4,7 +4,7 @@ module.exports = function(app) {
   })
 
   app.delete('/pagamentos/pagamento/:id', (req, res) => {
-    let pagamento = req.body
+    let pagamento = req.body['pagamento']
     pagamento.id = req.params.id
     pagamento.status = 'CANCELADO'
 
@@ -23,10 +23,10 @@ module.exports = function(app) {
   })
 
   app.put('/pagamentos/pagamento/:id', (req, res) => {
-    req.assert('forma_de_pagamento',
+    req.assert('pagamento.forma_de_pagamento',
       'Forma de pagamento é de preenchimento obrigatório'). notEmpty()
 
-    req.assert('valor',
+    req.assert('pagamento.valor',
       'Valor é de preenchimento obrigatório e deve ser um decimal')
         .notEmpty().isFloat()
 
@@ -38,7 +38,7 @@ module.exports = function(app) {
       return
     }
 
-    let pagamento = req.body
+    let pagamento = req.body['pagamento']
     pagamento.id = req.params.id
     pagamento.data = new Date
 
@@ -56,9 +56,9 @@ module.exports = function(app) {
   })
 
   app.post('/pagamentos/pagamento', (req, res) => {
-    req.assert('forma_de_pagamento',
+    req.assert('pagamento.forma_de_pagamento',
      'Forma de pagamento é de preenchimento obrigatório').notEmpty()
-    req.assert('valor',
+    req.assert('pagamento.valor',
      "Valor é de preenchimento obrigatório e deve ser um decimal")
       .notEmpty().isFloat()
 
@@ -70,7 +70,7 @@ module.exports = function(app) {
       return
     }
 
-    let pagamento = req.body
+    let pagamento = req.body['pagamento']
 
     pagamento.status = 'Criado'
     pagamento.data = new Date
@@ -83,28 +83,44 @@ module.exports = function(app) {
         res.status(500).send(erro)
       } else {
         pagamento.id = resultado.insertId
-        res.location('/pagamentos/pagamento/' + pagamento.id)
         console.log('Pagamento criado')
 
-        let response = {
-          dados_do_pagamento: pagamento,
-          links: [
-            {
-              href: `http://localhost:3000/pagamentos/pagamento/${pagamento.id}`,
-              rel: 'confirmar',
-              method: 'PUT'
-            },
-            {
-              href: `http://localhost:3000/pagamentos/pagamento/${pagamento.id}`,
-              rel: 'cancelar',
-              method: 'DELETE'
+        if (pagamento.forma_de_pagamento == 'cartao') {
+          let cartao = req.body['cartao']          
+          let clienteCartoes = new app.servicos.clienteCartoes()
+
+          clienteCartoes.autorizar(cartao, (exception, request, response, retorno) => {
+
+            if (exception) {
+              console.log(exception)
+              res.status(400).send(exception)
+              return
             }
+            res.status(201).json(retorno)
+            return
+          })
 
-          ]
+        } else {
+          res.location('/pagamentos/pagamento/' + pagamento.id)
+
+          let response = {
+            dados_do_pagamento: pagamento,
+            links: [
+              {
+                href: `http://localhost:3000/pagamentos/pagamento/${pagamento.id}`,
+                rel: 'confirmar',
+                method: 'PUT'
+              },
+              {
+                href: `http://localhost:3000/pagamentos/pagamento/${pagamento.id}`,
+                rel: 'cancelar',
+                method: 'DELETE'
+              }
+
+            ]
+          }
+          res.status(201).json(response)
         }
-
-
-        res.status(201).json(response)
       }
     })
   })
